State machine v2 tests
