+++ app/src/test/java/ru/whoame/recogniser/ui/base/mock/model/TestDomainEvent.kt
@@ -1,0 +1,7 @@
+package ru.whoame.recogniser.ui.base.mock.model
+
+enum class TestDomainEvent {
+
+    Default, DefaultWithSideEffect, FromUiHandlerWithSideEffect
+
+}
