+++ app/src/test/java/ru/whoame/recogniser/ui/base/mock/model/TestSideEffect.kt
@@ -1,0 +1,7 @@
+package ru.whoame.recogniser.ui.base.mock.model
+
+enum class TestSideEffect {
+
+    FromUiHandler, FromDomainHandler, FromUiAndDomainHandler
+
+}
