+++ app/src/test/java/ru/whoame/recogniser/ui/base/mock/model/TestState.kt
@@ -1,0 +1,4 @@
+package ru.whoame.recogniser.ui.base.mock.model
+
+// Dummy types for the base class generics
+data class TestState(val value: String = "Default")
