+++ app/src/test/java/ru/whoame/recogniser/ui/base/mock/model/TestUiEvent.kt
@@ -1,0 +1,7 @@
+package ru.whoame.recogniser.ui.base.mock.model
+
+enum class TestUiEvent {
+
+    Event, EventWithSideEffect, EventWithSideEffectInvokeDomainEventWithSideEffect
+
+}
