+++ app/src/test/java/ru/whoame/recogniser/ui/base/mock/model/TestUiState.kt
@@ -1,0 +1,3 @@
+package ru.whoame.recogniser.ui.base.mock.model
+
+data class TestUiState(val value: String)
