--- app/src/test/java/ru/whoame/recogniser/ui/viewmodel/handler/mock/TestDomainEvent.kt
@@ -1,7 +1,0 @@
-package ru.whoame.recogniser.ui.viewmodel.handler.mock
-
-enum class TestDomainEvent {
-
-    UpdateState, SideEffect, SideEffectLambda, UpdateStateWithSideEffect, TwoStateUpdatesWithDelay
-
-}
