--- app/src/test/java/ru/whoame/recogniser/ui/viewmodel/handler/mock/TestUiEvent.kt
@@ -1,13 +1,0 @@
-package ru.whoame.recogniser.ui.viewmodel.handler.mock
-
-enum class TestUiEvent {
-
-    UpdateState,
-    SideEffect,
-    SideEffectLambda,
-    UpdateStateWithSideEffect,
-    TwoStateUpdatesWithDelay,
-    DomainEvent,
-    DomainEventLambda,
-
-}
