--- app/src/test/java/ru/whoame/recogniser/ui/viewmodel/mock/handler/TestUiEventHandler.kt
@@ -1,36 +1,0 @@
-package ru.whoame.recogniser.ui.viewmodel.mock.handler
-
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.map
-import ru.whoame.recogniser.ui.viewmodel.handler.BaseUiEventHandler
-import ru.whoame.recogniser.ui.viewmodel.mock.model.*
-
-class TestUiEventHandler : BaseUiEventHandler<TestState, TestSideEffect, TestUiEvent, TestDomainEvent>() {
-
-    override fun handleEvent(event: TestUiEvent) = when (event) {
-        TestUiEvent.Event -> event()
-        TestUiEvent.EventWithSideEffect -> eventWithSideEffect()
-        TestUiEvent.EventWithSideEffectInvokeDomainEventWithSideEffect ->
-            eventWithSideEffectInvokeDomainEventWithSideEffect()
-    }
-
-    private fun event() = flow { emit(Unit) }
-        .map {
-            reduceState { state ->
-                state.copy(state.value + "_UiEvent")
-            }
-        }
-
-    private fun eventWithSideEffect() = event()
-        .map {
-            reduceSideEffect(TestSideEffect.FromUiHandler)
-        }
-
-    private fun eventWithSideEffectInvokeDomainEventWithSideEffect() = eventWithSideEffect()
-        .map {
-            reduceEvent {
-                TestDomainEvent.FromUiHandlerWithSideEffect
-            }
-        }
-
-}
