--- app/src/test/java/ru/whoame/recogniser/ui/viewmodel/mock/model/TestDomainEvent.kt
@@ -1,7 +1,0 @@
-package ru.whoame.recogniser.ui.viewmodel.mock.model
-
-enum class TestDomainEvent {
-
-    Default, DefaultWithSideEffect, FromUiHandlerWithSideEffect
-
-}
