--- app/src/test/java/ru/whoame/recogniser/ui/viewmodel/mock/model/TestState.kt
@@ -1,4 +1,0 @@
-package ru.whoame.recogniser.ui.viewmodel.mock.model
-
-// Dummy types for the base class generics
-data class TestState(val value: String = "Default")
