--- app/src/test/java/ru/whoame/recogniser/ui/viewmodel/mock/model/TestUiEvent.kt
@@ -1,7 +1,0 @@
-package ru.whoame.recogniser.ui.viewmodel.mock.model
-
-enum class TestUiEvent {
-
-    Event, EventWithSideEffect, EventWithSideEffectInvokeDomainEventWithSideEffect
-
-}
