--- app/src/test/java/ru/whoame/recogniser/ui/viewmodel/mock/model/TestUiState.kt
@@ -1,3 +1,0 @@
-package ru.whoame.recogniser.ui.viewmodel.mock.model
-
-data class TestUiState(val value: String)
