+++ state_machine/src/test/java/ru/whoame/state_machine/handler/mock/TestDomainEvent.kt
@@ -1,0 +1,7 @@
+package ru.whoame.state_machine.handler.mock
+
+enum class TestDomainEvent {
+
+    UpdateState, SideEffect, SideEffectLambda, UpdateStateWithSideEffect, TwoStateUpdatesWithDelay
+
+}
