+++ state_machine/src/test/java/ru/whoame/state_machine/handler/mock/TestUiEvent.kt
@@ -1,0 +1,13 @@
+package ru.whoame.state_machine.handler.mock
+
+enum class TestUiEvent {
+
+    UpdateState,
+    SideEffect,
+    SideEffectLambda,
+    UpdateStateWithSideEffect,
+    TwoStateUpdatesWithDelay,
+    DomainEvent,
+    DomainEventLambda,
+
+}
