+++ state_machine/src/test/java/ru/whoame/state_machine/mock/factory/TestStateFactory.kt
@@ -1,0 +1,11 @@
+package ru.whoame.state_machine.mock.factory
+
+import ru.whoame.state_machine.contract.BaseStateFactory
+import ru.whoame.state_machine.mock.model.TestState
+import ru.whoame.state_machine.mock.model.TestUiState
+
+class TestStateFactory : BaseStateFactory<TestState, TestUiState> {
+
+    override fun convert(state: TestState) = TestUiState(state.value)
+
+}
