+++ state_machine/src/test/java/ru/whoame/state_machine/mock/model/TestDomainEvent.kt
@@ -1,0 +1,7 @@
+package ru.whoame.state_machine.mock.model
+
+enum class TestDomainEvent {
+
+    Default, DefaultWithSideEffect, FromUiHandlerWithSideEffect
+
+}
