+++ state_machine/src/test/java/ru/whoame/state_machine/mock/model/TestState.kt
@@ -1,0 +1,4 @@
+package ru.whoame.state_machine.mock.model
+
+// Dummy types for the base class generics
+data class TestState(val value: String = "Default")
