+++ state_machine/src/test/java/ru/whoame/state_machine/mock/model/TestUiState.kt
@@ -1,0 +1,3 @@
+package ru.whoame.state_machine.mock.model
+
+data class TestUiState(val value: String)
